fix: Move math lib to separate crate
