+++ libs/math/Cargo.toml
@@ -1,0 +1,11 @@
+[package]
+name = "math"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+anchor-lang = "0.28.0"
+ruint = "1.3.0"
+num-traits = "0.2.14"
