+++ libs/math/src/lib.rs
@@ -1,3 +1,4 @@
+pub mod errors;
 pub mod safe_math;
 pub mod u128x128_math;
 pub mod utils_math;
