+++ programs/quarry/Cargo.toml
@@ -25,6 +25,7 @@ minter = {path = "../minter", features = ["cpi", "no-entrypoint"]}
 spl-math = { version = "0.1.0", features = ["no-entrypoint"] }
 u128 = "0.1.0"
 amm = {path = "../../libs/amm" }
+math = { workspace = true }
 ruint = "1.3.0"
